requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
importlib-metadata==5.0.0
Markdown==3.4.1
promise==2.3
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
zipp==3.10.0
